requirements.txt
django
django-environ
social-auth-app-django
django-ckeditor
django-extra-views
Pillow
geopy
cryptography
shippo
# book_manager
